Use apache commons pool to handle connections

+++ components/org.wso2.carbon.stream.processor.core/src/main/java/org/wso2/carbon/stream/processor/core/event/queue/EventListMapManager.java
@@ -28,13 +28,13 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentSkipListMap;
 
 public class EventListMapManager {
-    private static ConcurrentSkipListMap<Integer,QueuedEvent> eventListMap;
+    private static ConcurrentSkipListMap<Long,QueuedEvent> eventListMap;
 
     public EventListMapManager() {
     }
 
-    public static ConcurrentSkipListMap<Integer,QueuedEvent> initializeEventTreeMap() {
-        eventListMap = new ConcurrentSkipListMap<Integer,QueuedEvent>();
+    public static ConcurrentSkipListMap<Long,QueuedEvent> initializeEventTreeMap() {
+        eventListMap = new ConcurrentSkipListMap<Long, QueuedEvent>();
         return eventListMap;
     }
 
@@ -51,8 +51,8 @@ public class EventListMapManager {
                 Collection<List<Source>> sourceCollection = entry.getValue().getSiddhiAppRuntime().getSources();
                 for (List<Source> sources : sourceCollection) {
                     for (Source source : sources) {
-                        for(Map.Entry<Integer,QueuedEvent> listMapValue : eventListMap.entrySet()) {
-                            int key = listMapValue.getKey();
+                        for(Map.Entry<Long,QueuedEvent> listMapValue : eventListMap.entrySet()) {
+                            long key = listMapValue.getKey();
                             QueuedEvent queuedEvent = listMapValue.getValue();
                             if(queuedEvent.getSourceHandlerElementId().equals(source.getMapper().
                                     getHandler().getElementId()) &&
@@ -74,8 +74,8 @@ public class EventListMapManager {
                 String[] details = appDetail.split("__");//todo add constants
                 String appName = details[1].trim();
                 long timestamp = Long.valueOf(details[0].trim());
-                for(Map.Entry<Integer,QueuedEvent> treeMapValue : eventListMap.entrySet()) {
-                    int key = treeMapValue.getKey();
+                for(Map.Entry<Long,QueuedEvent> treeMapValue : eventListMap.entrySet()) {
+                    long key = treeMapValue.getKey();
                     QueuedEvent queuedEvent = treeMapValue.getValue();
                     if(queuedEvent.getSiddhiAppName().equals(appName) &&
                             timestamp >= queuedEvent.getEvent().getTimestamp()){
@@ -87,7 +87,7 @@ public class EventListMapManager {
         }
     }
 
-    public void addToTreeMap(int sequenceNum, QueuedEvent queuedEvent){
+    public void addToTreeMap(long sequenceNum, QueuedEvent queuedEvent){
         eventListMap.put(sequenceNum, queuedEvent);
     }
 }
